requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.18.0
streamlit-lottie>=0.0.5
streamlit-extras>=0.3.0
jupyter>=1.0.0
kaggle>=1.5.16
